requirements.txt
bcrypt==4.3.0
fastapi==0.115.12
PyJWT==2.10.1
SQLAlchemy==2.0.40
uvicorn==0.34.2
pytest==8.3.5